update expression iterators
