+++ kratos/containers/container_expression/expressions/expression.cpp
@@ -29,4 +29,85 @@ std::size_t Expression::GetFlattenedShapeSize() const
         [](const auto V1, const auto V2) { return V1 * V2; });
 }
 
+Expression::ExpressionIterator::ExpressionIterator(Expression::Pointer pExpression)
+    : mpExpression(pExpression),
+      mEntityIndex(0),
+      mEntityDataBeginIndex(0),
+      mComponentIndex(0),
+      mFlattenedShapeSize(pExpression->GetFlattenedShapeSize())
+{
+
+}
+
+Expression::ExpressionIterator::ExpressionIterator(const ExpressionIterator& rOther)
+    : mpExpression(rOther.mpExpression),
+      mEntityIndex(rOther.mEntityIndex),
+      mEntityDataBeginIndex(rOther.mEntityDataBeginIndex),
+      mComponentIndex(rOther.mComponentIndex),
+      mFlattenedShapeSize(rOther.mFlattenedShapeSize)
+{
+
+}
+
+Expression::Pointer Expression::ExpressionIterator::GetExpression() const
+{
+    return mpExpression;
+}
+
+double Expression::ExpressionIterator::operator*() const
+{
+    return mpExpression->Evaluate(mEntityIndex, mEntityDataBeginIndex, mComponentIndex);
+}
+
+bool Expression::ExpressionIterator::operator==(const ExpressionIterator& rOther) const
+{
+    return (&*mpExpression == &*rOther.mpExpression) && (mEntityIndex == rOther.mEntityIndex) && (mComponentIndex == rOther.mComponentIndex);
+}
+
+bool Expression::ExpressionIterator::operator!=(const ExpressionIterator& rOther) const
+{
+    return !this->operator==(rOther);
+}
+
+Expression::ExpressionIterator& Expression::ExpressionIterator::operator=(const ExpressionIterator& rOther)
+{
+    mpExpression = rOther.mpExpression;
+    mEntityIndex = rOther.mEntityIndex;
+    mEntityDataBeginIndex = rOther.mEntityDataBeginIndex;
+    mComponentIndex = rOther.mComponentIndex;
+    mFlattenedShapeSize = rOther.mFlattenedShapeSize;
+    return *this;
+}
+
+Expression::ExpressionIterator& Expression::ExpressionIterator::operator++()
+{
+    ++mComponentIndex;
+    if (mComponentIndex == mFlattenedShapeSize) {
+        mComponentIndex = 0;
+        ++mEntityIndex;
+        mEntityDataBeginIndex = mEntityIndex * mFlattenedShapeSize;
+    }
+    return *this;
+}
+
+Expression::ExpressionIterator Expression::ExpressionIterator::operator++(int)
+{
+    ExpressionIterator temp = *this;
+    ++*this;
+    return temp;
+}
+
+Expression::ExpressionIterator Expression::cbegin() const
+{
+    return ExpressionIterator(this);
+}
+
+Expression::ExpressionIterator Expression::cend() const
+{
+    ExpressionIterator result(this);
+    result.mEntityIndex = this->NumberOfEntities();
+    result.mEntityDataBeginIndex = result.mEntityIndex * result.mFlattenedShapeSize;
+    return result;
+}
+
 } // namespace Kratos
