+++ kratos/tests/cpp_tests/containers/test_expression_iterators.cpp
@@ -17,7 +17,6 @@
 
 // Project includes
 #include "testing/testing.h"
-#include "containers/container_expression/expressions/expression_iterator.h"
 #include "containers/container_expression/expressions/literal/literal_flat_expression.h"
 #include "containers/container_expression/expressions/binary/binary_expression.h"
 
@@ -42,9 +41,8 @@ KRATOS_TEST_CASE_IN_SUITE(ExpressionIteratorLazy, KratosCoreFastSuite) {
 
     auto p_binary_expression = BinaryExpression<BinaryOperations::Multiplication>::Create(p_literal_array_3_expression, p_literal_scalar_expression);
 
-    ExpressionIterator<Expression> itr(p_binary_expression);
     std::size_t local_index = 0;
-    std::for_each(itr.cbegin(), itr.cend(), [&p_binary_expression, &local_index](const auto Value) {
+    std::for_each(p_binary_expression->cbegin(), p_binary_expression->cend(), [&p_binary_expression, &local_index](const auto Value) {
         std::size_t entity_index = local_index / 3;
         std::size_t component_index = local_index % 3;
         std::size_t entity_data_begin_index = entity_index * 3;
@@ -59,25 +57,23 @@ KRATOS_TEST_CASE_IN_SUITE(ExpressionIteratorFlatChar, KratosCoreFastSuite) {
     auto p_literal_scalar_expression = LiteralFlatExpression<char>::Create(10, {});
     auto p_literal_array_3_expression = LiteralFlatExpression<char>::Create(10, {3});
 
-    ExpressionIterator<LiteralFlatExpression<char>> scalar_itr(p_literal_scalar_expression);
     char value = 0.0;
-    std::for_each(scalar_itr.begin(), scalar_itr.end(), [&value](auto& rValue) {
+    std::for_each(p_literal_scalar_expression->begin(), p_literal_scalar_expression->end(), [&value](auto& rValue) {
         rValue = value;
         value += 1;
     });
 
-    auto scalar_begin = scalar_itr.cbegin();
+    auto scalar_begin = p_literal_scalar_expression->cbegin();
     for (std::size_t i = 0; i < 10; ++i) {
         KRATOS_CHECK_EQUAL(*(scalar_begin + i), static_cast<char>(i));
     }
 
-    ExpressionIterator<LiteralFlatExpression<char>> array_3_itr(p_literal_array_3_expression);
-    std::for_each(array_3_itr.begin(), array_3_itr.end(), [&value](auto& rValue) {
+    std::for_each(p_literal_array_3_expression->begin(), p_literal_array_3_expression->end(), [&value](auto& rValue) {
         rValue = value;
         value += 1;
     });
 
-    auto vector_begin = array_3_itr.cbegin();
+    auto vector_begin = p_literal_array_3_expression->cbegin();
     for (std::size_t i = 0; i < 10; ++i) {
         KRATOS_CHECK_EQUAL(*(vector_begin + i), static_cast<char>(i + 10));
     }
@@ -87,25 +83,23 @@ KRATOS_TEST_CASE_IN_SUITE(ExpressionIteratorFlatInt, KratosCoreFastSuite) {
     auto p_literal_scalar_expression = LiteralFlatExpression<int>::Create(10, {});
     auto p_literal_array_3_expression = LiteralFlatExpression<int>::Create(10, {3});
 
-    ExpressionIterator<LiteralFlatExpression<int>> scalar_itr(p_literal_scalar_expression);
     int value = 0.0;
-    std::for_each(scalar_itr.begin(), scalar_itr.end(), [&value](auto& rValue) {
+    std::for_each(p_literal_scalar_expression->begin(), p_literal_scalar_expression->end(), [&value](auto& rValue) {
         rValue = value;
         value += 1;
     });
 
-    auto scalar_begin = scalar_itr.cbegin();
+    auto scalar_begin = p_literal_scalar_expression->cbegin();
     for (std::size_t i = 0; i < 10; ++i) {
         KRATOS_CHECK_EQUAL(*(scalar_begin + i), static_cast<int>(i));
     }
 
-    ExpressionIterator<LiteralFlatExpression<int>> array_3_itr(p_literal_array_3_expression);
-    std::for_each(array_3_itr.begin(), array_3_itr.end(), [&value](auto& rValue) {
+    std::for_each(p_literal_array_3_expression->begin(), p_literal_array_3_expression->end(), [&value](auto& rValue) {
         rValue = value;
         value += 1;
     });
 
-    auto vector_begin = array_3_itr.cbegin();
+    auto vector_begin = p_literal_array_3_expression->cbegin();
     for (std::size_t i = 0; i < 10; ++i) {
         KRATOS_CHECK_EQUAL(*(vector_begin + i), static_cast<int>(i + 10));
     }
@@ -115,25 +109,23 @@ KRATOS_TEST_CASE_IN_SUITE(ExpressionIteratorFlatDouble, KratosCoreFastSuite) {
     auto p_literal_scalar_expression = LiteralFlatExpression<double>::Create(10, {});
     auto p_literal_array_3_expression = LiteralFlatExpression<double>::Create(10, {3});
 
-    ExpressionIterator<LiteralFlatExpression<double>> scalar_itr(p_literal_scalar_expression);
     double value = 0.0;
-    std::for_each(scalar_itr.begin(), scalar_itr.end(), [&value](auto& rValue) {
+    std::for_each(p_literal_scalar_expression->begin(), p_literal_scalar_expression->end(), [&value](auto& rValue) {
         rValue = value;
         value += 1;
     });
 
-    auto scalar_begin = scalar_itr.cbegin();
+    auto scalar_begin = p_literal_scalar_expression->cbegin();
     for (std::size_t i = 0; i < 10; ++i) {
         KRATOS_CHECK_EQUAL(*(scalar_begin + i), static_cast<double>(i));
     }
 
-    ExpressionIterator<LiteralFlatExpression<double>> array_3_itr(p_literal_array_3_expression);
-    std::for_each(array_3_itr.begin(), array_3_itr.end(), [&value](auto& rValue) {
+    std::for_each(p_literal_array_3_expression->begin(), p_literal_array_3_expression->end(), [&value](auto& rValue) {
         rValue = value;
         value += 1;
     });
 
-    auto vector_begin = array_3_itr.cbegin();
+    auto vector_begin = p_literal_array_3_expression->cbegin();
     for (std::size_t i = 0; i < 10; ++i) {
         KRATOS_CHECK_EQUAL(*(vector_begin + i), static_cast<double>(i + 10));
     }
